+++ Database/Epoch/epochQuestDB.lua
@@ -746,5 +746,10 @@ epochQuestData[28748] = {"Pushing Forward",{{944}},{{944}},nil,5,nil,nil,nil,nil
 epochQuestData[28749] = {"A Refugee's Quandary",nil,nil,nil,3,nil,nil,nil,nil,nil,nil,nil,nil,nil,nil,nil,85,nil,nil,nil,nil,nil,nil,0,nil,nil,nil,nil,nil,nil}
 epochQuestData[28756] = {"Return to Mother",{{944}},{{45999}},nil,5,nil,nil,nil,nil,nil,nil,nil,nil,nil,nil,nil,85,nil,nil,nil,nil,nil,nil,0,nil,nil,nil,nil,nil,nil}
 
+-- Issue #38: Additional quests
+epochQuestData[26289] = {"Renegade Naga",{{45190}},{{7877}},nil,42,nil,nil,nil,nil,nil,nil,nil,nil,nil,nil,nil,85,nil,nil,nil,nil,nil,nil,0,nil,nil,nil,nil,nil,nil}
+epochQuestData[26594] = {"WANTED: Scorchmaw",nil,nil,nil,45,nil,nil,nil,nil,nil,nil,nil,nil,nil,nil,nil,85,nil,nil,nil,nil,nil,nil,0,nil,nil,nil,nil,nil,nil}
+epochQuestData[26863] = {"The Shadowforge Librarian",nil,nil,nil,45,nil,nil,nil,nil,nil,nil,nil,nil,nil,nil,nil,85,nil,nil,nil,nil,nil,nil,0,nil,nil,nil,nil,nil,nil}
+
 -- Stage the Epoch questData for later merge during compilation
 QuestieDB._epochQuestData = epochQuestData